% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_initialization.R
\name{PCA_ordering}
\alias{PCA_ordering}
\title{Principal-component ordering}
\usage{
PCA_ordering(X, center = TRUE, scale = FALSE, scale01 = TRUE, component = 1L)
}
\arguments{
\item{X}{numeric matrix (n x D).}

\item{center}{logical.}

\item{scale}{logical.}

\item{scale01}{logical; if TRUE, rescale returned t to [0,1].}

\item{component}{Integer principal-component index to use as the ordering
score. Defaults to 1.}
}
\value{
list(t, keep_idx)
}
\description{
Principal-component ordering
}
