% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPCurver-package.R
\docType{package}
\name{MPCurver}
\alias{MPCurver-package}
\alias{MPCurver}
\title{MPCurver package}
\description{
Estimate latent sample orderings, smooth feature trajectories, and
trajectory uncertainty using Gaussian measurement models and random-walk
priors. Multi-ordering models also estimate the probability that each
feature follows each ordering, using structural variational inference.
Start with \code{\link{fit_mpcurve}} and explore the results with
\code{\link{summary.mpcurve}} and \code{\link{plot.mpcurve}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aguerozz.github.io/MPCurver/}
  \item \url{https://github.com/AgueroZZ/MPCurver}
  \item Report bugs at \url{https://github.com/AgueroZZ/MPCurver/issues}
}

}
\author{
\strong{Maintainer}: Ziang Zhang \email{aguero.zhang@gmail.com}

}
\keyword{internal}
